requirements.txt
atoma==0.0.17
feedgen==0.9.0
httpx==0.25.2
python-dotenv==1.0.0
ratelimit==2.2.1
rich==13.7.0
tendo==0.3.0
.
